The user has the right to choose a visible letter.
If the user chooses a visible letter as a guess, they receive an alarm, explaining that their guess is visible, therefore it's not correct. However the user has the right to keep it as one of their guesses. Also by choosing another visible letter and this time clicking on the Cancel button, the first visible letter won't disappear.