Changing the color of the selected block.
When the user clicks on a block, not only the value of this block is considered as of of their guesses, but also the color of this block changes to gray. Of course if the user clicks on an already selected block, its value is removed from their guesses and its gray color disappears.