Fixing some bugs.
There were some issues with the "Reset the game" button: By reseing the game, the user wouldn't retrieve the chance of the helping buttons (seeing all the cards for 3 seconds and finding the match). Also the resolved cards would reveal their picture when the "See all the cards for 3 seconds" button would be clicked.